third_party/WebKit/LayoutTests/fast/forms/text/input-set-value-on-blur.html

<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<textarea id="text">text1</textarea>
<input id="input" value="text2" />
<script>
description("This test confirms that changing an input's value on blur event does not cause re-focus.");

input.select();
input.addEventListener('blur', function () {
    input.value = "new value";
});
text.focus();

shouldBe('document.activeElement', 'text')

</script>
</body>
</html>
</body>
