third_party/blink/web_tests/fast/forms/text/input-set-value-on-blur-expected.txt

This test confirms that changing an input's value on blur event does not cause re-focus.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.activeElement is text
PASS successfullyParsed is true

TEST COMPLETE

 
